Make the Origin available even without iterators
The Origin can now be extracted directly from `siginfo_t`, even without
using or enabling the iterator.
