+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "signal-hook"
-version = "0.3.1"
+version = "0.3.2"
 authors = [
     "Michal 'vorner' Vaner <vorner@vorner.cz>",
     "Thomas Himmelstoss <thimm@posteo.de>",
@@ -35,7 +35,7 @@ members = [
 [dependencies]
 libc = "^0.2"
 signal-hook-registry = { version = "^1.3", path = "signal-hook-registry" }
-signal-hook-sys = { version = "^0.1", path = "signal-hook-sys", optional = true }
+signal-hook-sys = { version = "^0.1.1", path = "signal-hook-sys", optional = true }
 
 [dev-dependencies]
 serial_test = "^0.5"
